Added fixing

+++ Company.Delivery.Database/Repos/WaybillRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Company.Delivery.Core;
+using Company.Delivery.Domain;
 using Company.Delivery.Domain.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
@@ -13,21 +14,18 @@ namespace Company.Delivery.Database.Repos
         }
         public async Task<Waybill> CreateAsync(Waybill waybill, CancellationToken cancellationToken)
         {
-            try
-            {
-                await _dbContext.Waybills.AddAsync(waybill, cancellationToken);
-                await _dbContext.SaveChangesAsync(cancellationToken);
-                return waybill;
-            }
-            catch (Exception ex)
-            {
-                var t = ex.Message;
-                throw;
-            }
+            await _dbContext.Waybills.AddAsync(waybill, cancellationToken);
+            await _dbContext.SaveChangesAsync(cancellationToken);
+            return waybill;
         }
         public async Task DeleteByIdAsync(Guid id, CancellationToken cancellationToken)
         {
             var waybill = await _dbContext.Waybills.FindAsync(id, cancellationToken);
+            if (waybill is null)
+            {
+                throw new EntityNotFoundException();
+            }
+
             _dbContext.Waybills.Remove(waybill);
             await _dbContext.SaveChangesAsync(cancellationToken);
         }
